requirements.txt
asgiref==3.2.10

attrs==20.2.0
autopep8==1.5.4
beautifulsoup4==4.9.1
boto3==1.16.10
botocore==1.19.10
dj-database-url==0.5.0
Django==3.1
django-bootstrap4==2.2.0
django-crispy-forms==1.9.2
django-forms-bootstrap==3.0.0
django-haystack==3.0b2
django-js-asset==1.2.2
django-machina==1.1.4
django-mptt==0.11.0
django-password-reset==2.0
django-phone-field==1.8.1
django-storages==1.10.1
django-tinymce==3.0.2
Django-tinymce-filebrowser==0.2.1
django-widget-tweaks==1.4.8
factory-boy==3.0.1
Faker==4.1.3
gunicorn==20.0.4
html5lib==1.1
importlib-metadata==1.7.0
inflection==0.5.1
iniconfig==1.0.1
jmespath==0.10.0
markdown2==2.4.0
more-itertools==8.5.0
packaging==20.4
Pillow
pluggy==0.13.1
psycopg2-binary
py==1.9.0
pycodestyle==2.6.0
pyparsing==2.4.7
PyPDF2==1.26.0
pytest==6.0.2
pytest-factoryboy==2.0.3
python-dateutil==2.8.1
pytz==2020.1
reportlab==3.5.48
s3transfer==0.3.3
six==1.15.0
sorl-thumbnail==12.6.3
soupsieve==2.0.1
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.1
urllib3==1.25.11
webencodings==0.5.1
whitenoise==5.2.0
xhtml2pdf==0.2.4
zipp==3.2.0
django-taggit==1.3.0
